Fix: 회원가입 페이지 url 방식으로 수정
