+++ webapp/src/pages/SignUp/IdPassword.jsx
@@ -1,0 +1,74 @@
+import React from 'react';
+import { useForm } from 'react-hook-form';
+import { useNavigate } from 'react-router-dom';
+import { handleFetcher } from 'utils';
+import authApi from 'api/auth';
+
+export default function IdPassword() {
+  const navigate = useNavigate();
+  const {
+    register,
+    setError,
+    formState: { errors },
+  } = useForm({
+    defaultValues: {},
+  });
+  const { handleSubmit } = useForm({
+    defaultValues: {},
+  });
+  const onValid = async (submitData) => {
+    const { email, nickname, password, verifiedPassword } = submitData;
+    if (password !== verifiedPassword) {
+      setError('verifiedPassword', { message: 'Password is not same' }, { shouldFocus: true });
+    }
+    const signUpInfo = {
+      email,
+      name: nickname,
+      pwd: password,
+    };
+    // TODO: input validation 추가해야 함.
+    const { value, error, isError } = await handleFetcher(authApi.POST_SIGN_UP, signUpInfo);
+    if (isError) {
+      console.log(error);
+      return;
+    }
+    navigate('/login');
+  };
+  return (
+    <div>
+      <form style={{ display: 'flex', flexDirection: 'column' }} onSubmit={handleSubmit(onValid)}>
+        <input
+          {...register('email', {
+            required: 'Email is required',
+            pattern: {
+              value: /^[_a-z0-9-]+(.[_a-z0-9-]+)*@(?:\w+\.)+\w+$/i,
+              message: '이메일 형식으로 입력해주세요.',
+            },
+          })}
+          placeholder="email"
+        />
+        <span>{errors?.email?.message}</span>
+        <input
+          {...register('password', {
+            required: '4자리 이상 비밀번호를 입력해주세요.',
+            minLength: 4,
+            pattern: {
+              value: /(?=.*[0-9])(?=.*[a-z])(?=.*[!@#$%^&+=])(?=\S+$).{8,20}/,
+              message:
+                '8자 이상 20자 이하, 숫자 한개이상 특수문자 한개이상 영어 한개이상 포함 공백 불가',
+            },
+          })}
+          placeholder="password"
+        />
+        <span>{errors?.password?.message}</span>
+        <input
+          {...register('verifiedPassword', {
+            required: '비밀번호가 일치하지 않습니다.',
+          })}
+          placeholder="verifiedPassword"
+        />
+        <span>{errors?.verifiedPassword?.message}</span>
+      </form>
+    </div>
+  );
+}
