+++ webapp/src/pages/SignUp/SessionJob.jsx
@@ -1,6 +1,10 @@
+import useInput from 'hooks/useInput';
 import React from 'react';
+import { hopeSessionOption } from 'constant';
 
-export default function nickname() {
+export default function SessionJob() {
+  const [hopeSession, onHopeSessionChange] = useInput('무관');
+  const [userJob, onJobChange] = useInput('');
   return (
     <>
       <span>희망 작업 기간</span>
