+++ webapp/src/pages/SignUp/Skill.jsx
@@ -1,6 +1,13 @@
-import React from 'react';
+import React, { useCallback, useState } from 'react';
+import { skillOptions } from 'constant';
 
 export default function Skill() {
+  const [userSkill, setUserSkill] = useState('');
+  const [selectedSkills, setSelectedSkills] = useState([]);
+  const onSkillChange = useCallback((e) => {
+    setUserSkill(e.target.value);
+    setSelectedSkills((prev) => [...prev, e.target.value]);
+  }, []);
   return (
     <>
       <span>선택한 기술 스킬: {selectedSkills.join(', ')}</span>
